requirements.txt
botocore==1.24.3
dnspython==2.2.0
pymongo==4.1.1
pyTelegramBotAPI==4.4.0
s3transfer==0.5.1
telegram==0.0.1
python-dotenv==0.20.0
boto3==1.24.1
requests==2.27.1
